pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agri-insurance-engine"
version = "0.1.0"
description = "Parametric micro-insurance trigger engine for Indian farmers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "geoalchemy2>=0.15.0",
    "redis>=5.0.0",

    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",

    # Big Data
    "dask[array,distributed]>=2024.4.0",
    "pyspark>=3.5.0",

    # Geospatial
    "gdal>=3.8.0",
    "rasterio>=1.3.0",
    "rasterstats>=0.19.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "fiona>=1.9.0",

    # Task queue
    "celery[redis]>=5.3.0",

    # MQTT
    "paho-mqtt>=2.0.0",

    # Security / Encryption
    "cryptography>=42.0.0",
    "python-jose[cryptography]>=3.3.0",

    # Dashboard
    "streamlit>=1.34.0",
    "folium>=0.16.0",
    "plotly>=5.22.0",
    "streamlit-folium>=0.20.0",

    # Utilities
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "bandit>=1.7.0",
    "factory-boy>=3.3.0",
    "Faker>=25.0.0",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "S", "UP"]
ignore = ["S101"]  # allow assert in tests

[tool.black]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
